Split Entry into its own file
